module1/wallets/requirements.txt
aws-cdk-lib==2.79.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.79.1a0
cdk-nag==2.27.1
